Move export out of old crypto to api with re-export
