+++ src/crypto-api/keybackup.ts
@@ -15,6 +15,7 @@ limitations under the License.
 */
 
 import { ISigned } from "../@types/signed";
+import { IEncryptedPayload } from "../crypto/aes";
 
 export interface Curve25519AuthData {
     public_key: string;
@@ -67,4 +68,22 @@ export interface BackupTrustInfo {
 export interface KeyBackupCheck {
     backupInfo: KeyBackupInfo;
     trustInfo: BackupTrustInfo;
+}
+
+export interface Curve25519SessionData {
+    ciphertext: string;
+    ephemeral: string;
+    mac: string;
+}
+
+/* eslint-disable camelcase */
+export interface KeyBackupSession<T = Curve25519SessionData | IEncryptedPayload> {
+    first_message_index: number;
+    forwarded_count: number;
+    is_verified: boolean;
+    session_data: T;
+}
+
+export interface KeyBackupRoomSessions {
+    [sessionId: string]: KeyBackupSession;
 }
